fix: move Developer Mode toggle from Appearance to Profile section
