+++ src/frontend/components/settings/AppearanceSettings.tsx
@@ -1,6 +1,5 @@
 import { useAppDispatch, useAppSelector } from '../../redux/hooks';
-import { selectTheme, setTheme, selectDeveloperMode, setDeveloperMode } from '../../redux/slices/userSettings';
-import { isPrivilegedUser } from '../../lib/role-utils';
+import { selectTheme, setTheme } from '../../redux/slices/userSettings';
 import { Sun, Moon, Monitor } from 'lucide-react';
 
 type ThemeOption = 'light' | 'dark';
@@ -13,7 +12,6 @@ const THEME_OPTIONS: { value: ThemeOption; label: string; icon: typeof Sun; desc
 export function AppearanceSettings() {
   const dispatch = useAppDispatch();
   const currentTheme = useAppSelector(selectTheme);
-  const developerMode = useAppSelector(selectDeveloperMode);
 
   return (
     <div className="space-y-6">
@@ -53,31 +51,6 @@ export function AppearanceSettings() {
           })}
         </div>
       </div>
-
-      {isPrivilegedUser() && (
-        <div>
-          <h3 className="text-sm font-medium text-foreground mb-1">Developer Mode</h3>
-          <p className="text-xs text-muted-foreground mb-3">
-            Show the Developer tab with advanced tools and agent evaluation.
-          </p>
-          <button
-            type="button"
-            role="switch"
-            aria-checked={developerMode}
-            aria-label="Enable Developer Mode"
-            onClick={() => dispatch(setDeveloperMode(!developerMode))}
-            className={`relative inline-flex h-5 w-9 items-center rounded-full transition-colors cursor-pointer ${
-              developerMode ? 'bg-primary' : 'bg-muted'
-            }`}
-          >
-            <span
-              className={`inline-block h-3.5 w-3.5 transform rounded-full bg-white shadow transition-transform ${
-                developerMode ? 'translate-x-4' : 'translate-x-1'
-              }`}
-            />
-          </button>
-        </div>
-      )}
 
       <div>
         <h3 className="text-sm font-medium text-foreground mb-1">Interface Density</h3>
